Splitting normalization files (#449)
* Splitting the normalization files.

* Splitting groupnorm and instancenorm.

* Forgot to delete old file.

* Fixed pep8

* Fixing imports by adding an init file.
